pyproject.toml
[tool.poetry]
name = "telegram-messages-manager"
version = "0.1.0"
description = ""
authors = ["Matheus Albino Brunhara <matheusalb3213@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.30.3"
cryptg = "^0.4.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
